Move modals into shared folder
- Fix paths in working files
