+++ src/components/posts/share/SpaceShareLink.tsx
@@ -2,10 +2,10 @@ import React, { useState } from 'react'
 import { PostWithSomeDetails } from '@subsocial/types/dto';
 import { PostExtension } from '@subsocial/types/substrate/classes';
 import { EditOutlined } from '@ant-design/icons';
-import { ShareModal } from '../ShareModal'
-import { isRegularPost } from '../view-post';
-import { IconWithLabel } from '../../utils';
-import { useAuth } from '../../auth/AuthContext';
+import { ShareModal } from 'src/components/posts/modals/ShareModal'
+import { isRegularPost } from 'src/components/posts/view-post';
+import { IconWithLabel } from 'src/components/utils';
+import { useAuth } from 'src/components/auth/AuthContext';
 
 type Props = {
   postDetails: PostWithSomeDetails
